backend/pyproject.toml
[project]
name = "farm-automation-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "manuEcho360",email = "mmohan@echo360.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
